vpn-service/src/users_auth/serizlizers.py
from rest_framework import serializers
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from users.models import UserModel
from users.serializers import UserSerializer
from src.users.serizalizer import UserSerializer
from src.users.models import UserModel

class UserSignUpSerializer(serializers.ModelSerializer):
    """Serializer for user registration"""

    class Meta:
        model = User
        fields = ("email", "password", "first_name", "last_name")
        extra_kwargs = {"password": {"write_only": True}}

    def create(self, validated_data):
        user = super().create(validated_data)
        user.set_password(validated_data["password"])
        user.save()
        return user


class TokenSerializer(TokenObtainPairSerializer):
    """
    Serializer for obtaining and validating user tokens.
    """

    def validate(self, attrs):
        data = super().validate(attrs)
        data["user"] = UserSerializer(self.user).data
        return data