vpn-service/pyproject.toml
[tool.poetry]
name = "vpn-service"
version = "0.1.0"
description = ""
authors = ["Alex <124603897+OleksandrBrusyltsev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
